requirements.txt
mypy-extensions==1.0.0
mysql-connector-python==8.0.33
PyMySQL==1.1.0
python-dateutil==2.8.2
pytz==2023.3
tabulate==0.9.0
datetime==5.1
python-dotenv==1.0.0